feat: add TVM FFI support

+++ include/omniback/builtin/cat_split.hpp
@@ -15,9 +15,10 @@ class CatSplit final : public Container {
   virtual void impl_forward(const std::vector<dict>&) override;
 
  private:
-  virtual std::vector<size_t> set_init_order(size_t max_range) const override;
+  virtual std::vector<uint32_t> set_init_order(
+      uint32_t max_range) const override;
 
-  std::pair<size_t, size_t> update_min_max(
+  std::pair<uint32_t, uint32_t> update_min_max(
       const std::vector<Backend*>& depends) override;
 };
