+++ include/omniback/builtin/generate_backend.hpp
@@ -32,11 +32,11 @@ class GenerateBackend : public Backend {
   void impl_forward(const std::vector<dict>& inputs) override {
     proxy_backend_->forward(inputs);
   }
-  [[nodiscard]] virtual size_t impl_max() const override {
+  [[nodiscard]] virtual uint32_t impl_max() const override {
     return proxy_backend_->max();
   }
 
-  [[nodiscard]] virtual size_t impl_min() const override {
+  [[nodiscard]] virtual uint32_t impl_min() const override {
     return proxy_backend_->min();
   }
