+++ include/omniback/builtin/result_queue.hpp
@@ -60,7 +60,7 @@ class Send : public Backend {
  protected:
   // std::string target_name_;
   Queue* queue_{nullptr};
-  size_t queue_max_{std::numeric_limits<size_t>::max()};
+  size_t queue_max_{std::numeric_limits<uint32_t>::max()};
 };
 
 class Observer : public Send {
